Удаляет обработчики и прибирает мусор
